Added Input Command, updated tests.
